requirements.txt
fastapi
uvicorn[standard]
pydantic
PyYAML
pandas
plotly
dash
dash-bootstrap-components
joblib
httpx
python-multipart
loguru